chore: add support for mocking task for tests
With introduction of `transaction/2` it became harder to test sagas
wrapped into a database transaction. The issue is that once you have
a test which spawns a new process which needs to talk to a database
within a transaction, this process cannot obtain a database connection.

ExUnit process -> ExUnit transaction -> Sagax transaction -> Task.async
-> SQL query from Sagax stage -> error!

This somewhat hacky way of mocking Task module should fix the issue.